function-source/requirements.txt
# Function dependencies, for example:
# package>=version
line-bot-sdk
openai
yahoo_fin
requests_html
beautifulsoup4